Refactor for class methods rather than raw client
The documentation for ruby-trello[1] describes using class methods than
using the raw client.

I used the raw client initially since it felt easier to understand the
Trello API, but I've since decided it would be nicer to use class
methods.

This refactors the `Squiddy::Trello::Checklist` class to use these
methods.

Since we have similar namespacing, I've added to anchor the the
ruby-trello library using `::Trello` when referencing the classes.

This also adds the configuration to the top level using an initializer.

[1] https://github.com/jeremytregunna/ruby-trello