Add TrelloDependabot
This adds TrelloDependabot, which creates a card in the specified list
name from the parameter `list_create_name`, giving it the same name and
body as the pull request, along with a link to the PR.

When the PR is closed, then the card is moved to the `list_done_name`
column.

The `trello` library made it straight forward to achieve this behaviour.